refactor(app): remaining desktop components
