+++ app/src/organisms/Desktop/Devices/ErrorRecoveryBanner/__tests__/useErrorRecoveryBanner.test.ts
