+++ app/src/organisms/Desktop/ProtocolTimelineScrubber/PipetteVisuals.tsx
