+++ app/src/organisms/Desktop/ProtocolTimelineScrubber/utils.ts
