+++ app/src/pages/Desktop/Protocols/ProtocolDetails/ProtocolTimeline.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 import { Icon, Box, SPACING } from '@opentrons/components'
 import { fetchProtocols, getStoredProtocol } from '/app/redux/protocol-storage'
-import { ProtocolTimelineScrubber } from '/app/organisms/ProtocolTimelineScrubber'
+import { ProtocolTimelineScrubber } from '/app/organisms/Desktop/ProtocolTimelineScrubber'
 
 import type { Dispatch, State } from '/app/redux/types'
 import type { DesktopRouteParams } from '/app/App/types'
